Clean up test file naming and consolidate test structure
- Rename test files to remove descriptive adjectives/adverbs
- Consolidate to one main test file per source module plus focused basic tests
- test_config.py -> test_config_basic.py (core functionality)
- test_config_extended.py -> test_config.py (comprehensive testing)
- test_bot.py -> test_bot_basic.py (specific functions)
- test_bot_extended.py -> test_bot.py (comprehensive bot testing)
- All linters pass (black, ruff)
- All 408 tests pass
