+++ .github/workflows/combine-prs.yml
@@ -22,12 +22,16 @@ on:
         required: true
         default: 'nocombine'
 
-# A workflow run is made up of one or more jobs that can run sequentially or in parallel
+permissions:
+  contents: write
+  pull-requests: write
+
 jobs:
   # This workflow contains a single job called "combine-prs"
   combine-prs:
     # The type of runner that the job will run on
     runs-on: ubuntu-latest
+    timeout-minutes: 10  # Prevent hanging workflows
 
     # Steps represent a sequence of tasks that will be executed as part of the job
     steps:
@@ -39,7 +43,8 @@ jobs:
           script: |
             const pulls = await github.paginate('GET /repos/:owner/:repo/pulls', {
               owner: context.repo.owner,
-              repo: context.repo.repo
+              repo: context.repo.repo,
+              state: 'open'  // Only get open PRs for better performance
             });
             let branchesAndPRStrings = [];
             let baseBranch = null;
