+++ .github/workflows/docker.yml
@@ -14,6 +14,7 @@ env:
 jobs:
   build:
     runs-on: ubuntu-latest
+    timeout-minutes: 15  # Prevent hanging builds
     
     permissions:
       contents: read
@@ -21,6 +22,13 @@ jobs:
 
     steps:
     - uses: actions/checkout@v4
+    
+    - name: Set up Docker Buildx
+      uses: docker/setup-buildx-action@v3
+      with:
+        driver-opts: |
+          image=moby/buildkit:v0.12.0
+        platforms: linux/amd64,linux/arm64  # Multi-platform builds
     
     - name: Log in to the Container registry
       uses: docker/login-action@v3
@@ -37,11 +45,20 @@ jobs:
         tags: |
           type=raw,value=latest,event=branch
           type=ref,event=tag
+        labels: |
+          org.opencontainers.image.source=${{ github.server_url }}/${{ github.repository }}
+          org.opencontainers.image.revision=${{ github.sha }}
+          org.opencontainers.image.version=${{ github.ref_name }}
 
     - name: Build and push Docker image
       uses: docker/build-push-action@v5
       with:
         context: .
+        platforms: linux/amd64,linux/arm64
         push: true
         tags: ${{ steps.meta.outputs.tags }}
         labels: ${{ steps.meta.outputs.labels }}
+        cache-from: type=gha  # Use GitHub Actions cache
+        cache-to: type=gha,mode=max  # Cache layers for faster builds
+        build-args: |
+          BUILDKIT_INLINE_CACHE=1
